opentelemetry: dissolve api module in favor of top-level modules (#412)
This makes the source tree less nested and makes the internal API and
filesystem more similar to the public API exposed by the crate.
