+++ opentelemetry/src/global/error_handler.rs
@@ -1,9 +1,36 @@
-use crate::api::Error;
+use std::sync::PoisonError;
 use std::sync::RwLock;
+
+#[cfg(feature = "metrics")]
+use crate::metrics::MetricsError;
+#[cfg(feature = "trace")]
+use crate::trace::TraceError;
 
 lazy_static::lazy_static! {
     /// The global error handler.
     static ref GLOBAL_ERROR_HANDLER: RwLock<Option<ErrorHandler>> = RwLock::new(None);
+}
+
+/// Wrapper for error from both tracing and metrics part of open telemetry.
+#[derive(thiserror::Error, Debug)]
+#[non_exhaustive]
+pub enum Error {
+    #[cfg(feature = "trace")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "trace")))]
+    #[error(transparent)]
+    Trace(#[from] TraceError),
+    #[cfg(feature = "metrics")]
+    #[cfg_attr(docsrs, doc(cfg(feature = "metrics")))]
+    #[error(transparent)]
+    Metric(#[from] MetricsError),
+    #[error("{0}")]
+    Other(String),
+}
+
+impl<T> From<PoisonError<T>> for Error {
+    fn from(err: PoisonError<T>) -> Self {
+        Error::Other(err.to_string())
+    }
 }
 
 struct ErrorHandler(Box<dyn Fn(Error) + Send + Sync>);
