+++ opentelemetry/src/labels/encoder.rs
