+++ opentelemetry/src/metrics/counter.rs
