+++ opentelemetry/src/metrics/noop.rs
