+++ opentelemetry/src/metrics/number.rs
