+++ opentelemetry/src/trace/event.rs
