+++ opentelemetry/src/trace/link.rs
