+++ opentelemetry/src/trace/noop.rs
