+++ opentelemetry/src/trace/span.rs
